requirements.txt
-ensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.0.0-py3-none-any.whl
absl-py==0.9.0
altair==4.1.0
appdirs==1.4.3
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
backcall==0.1.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
blis==0.4.1
boto3==1.16.1
botocore==1.19.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
emoji==0.6.0
entrypoints==0.3
enum-compat==0.0.3
Flask==1.1.2
Flask-Cors==3.0.8
gast==0.3.3
geojson==2.5.0
gitdb==4.0.5
GitPython==3.1.9
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
imutils==0.5.3
inline==0.0.1
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonify==0.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.2
nbclient==0.5.1
nbconvert==5.6.1
nbformat==5.0.6
nest-asyncio==1.4.1
nltk==3.5
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
pandas-datareader==0.8.1
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.5.0b1
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
smmap==3.0.4
spacy==2.3.2
srsly==1.0.2
streamlit==0.69.2
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
Theano @ git+git://github.com/Theano/Theano.git@eb6a4125c4f5617e74b10503afc3f334f17cf545
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tqdm==4.48.2
traitlets==5.0.5
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.25.10
validators==0.18.1
wasabi==0.7.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.3.1
